# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for API e-mosque endpoints",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/zakat-fitrah": {
            "get": {
                "summary": "Read Zakat data",
                "description": "Read Zakat data.",
                "operationId": "readZakat",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of Zakat data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the data retrieval",
                                            "type": "string",
                                            "example": "Data Berhasil ditemukan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Zakat data not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the data retrieval failure",
                                            "type": "string",
                                            "example": "Data Zakat Tidak ditemukan"
                                        },
                                        "data": {
                                            "description": "Additional error data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the data retrieval failure",
                                            "type": "string",
                                            "example": "Gagal Menajalankan Request"
                                        },
                                        "data": {
                                            "description": "Additional error data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create new Zakat data",
                "description": "Create new Zakat data.",
                "operationId": "createZakat",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_pezakat": {
                                        "description": "Name of the zakat giver",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "jumlah_zakat": {
                                        "description": "Amount of zakat",
                                        "type": "integer",
                                        "example": 100000
                                    },
                                    "zakat_jenis_id": {
                                        "description": "ID of the zakat type",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful creation of Zakat data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the data creation",
                                            "type": "string",
                                            "example": "Data Zakat Berhasil Ditambahkan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the data creation failure",
                                            "type": "string",
                                            "example": "Gagal menjalankan request."
                                        },
                                        "data": {
                                            "description": "Additional error data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}